requirements.txt
flask
flask_session
pyyaml
requests
wbib==0.3.11
wdcuration
wikidata2df